search offering user story modification with creation of SkillManager and CategoryManager
